+++ src/lib_shell/prevalidation.ml
@@ -64,14 +64,6 @@ module type T = sig
 
   val apply_operation : t -> operation -> result Lwt.t
 
-  type status = {
-    applied_operations : (operation * Proto.operation_receipt) list;
-    block_result : Tezos_protocol_environment.validation_result;
-    block_metadata : Proto.block_header_metadata;
-  }
-
-  val status : t -> status tzresult Lwt.t
-
   val validation_state : t -> Proto.validation_state
 
   val pp_result : Format.formatter -> result -> unit
@@ -210,16 +202,6 @@ module Make (Proto : Tezos_protocol_environment.PROTOCOL) :
           | `Permanent -> Refused errors
           | `Temporary -> Branch_delayed errors)
 
-  type status = {
-    applied_operations : (operation * Proto.operation_receipt) list;
-    block_result : Tezos_protocol_environment.validation_result;
-    block_metadata : Proto.block_header_metadata;
-  }
-
-  let status pv =
-    Proto.finalize_block pv.state >>=? fun (block_result, block_metadata) ->
-    return {block_metadata; block_result; applied_operations = pv.applied}
-
   let validation_state {state; _} = state
 
   let pp_result ppf =
@@ -322,35 +304,39 @@ let preapply chain_store ~user_activated_upgrades
            Operation_list_hash.compute (List.map fst r.Preapply_result.applied))
          validation_result_list_rev)
   in
-  Prevalidation.status validation_state >>=? fun {block_result; _} ->
   let pred_shell_header = Store.Block.shell_header predecessor in
   let level = Int32.succ pred_shell_header.level in
+  let pred_block_hash = Store.Block.hash predecessor in
+  let shell_header : Block_header.shell_header =
+    {
+      level;
+      proto_level = pred_shell_header.proto_level;
+      predecessor = pred_block_hash;
+      timestamp;
+      validation_passes;
+      operations_hash;
+      context = Context_hash.zero (* place holder *);
+      fitness = [];
+    }
+  in
+  Proto.finalize_block (Prevalidation.validation_state validation_state)
+  >>=? fun (block_result, _) ->
   Block_validation.may_patch_protocol
     ~user_activated_upgrades
     ~user_activated_protocol_overrides
     ~level
     block_result
   >>= fun {fitness; context; message; _} ->
-  Store.Block.protocol_hash chain_store predecessor >>=? fun pred_protocol ->
   let context = Shell_context.unwrap_disk_context context in
   Context.get_protocol context >>= fun protocol ->
+  Store.Block.protocol_hash chain_store predecessor >>=? fun pred_protocol ->
   let proto_level =
     if Protocol_hash.equal protocol pred_protocol then
       pred_shell_header.proto_level
     else (pred_shell_header.proto_level + 1) mod 256
   in
-  let pred_block_hash = Store.Block.hash predecessor in
   let shell_header : Block_header.shell_header =
-    {
-      level;
-      proto_level;
-      predecessor = pred_block_hash;
-      timestamp;
-      validation_passes;
-      operations_hash;
-      fitness;
-      context = Context_hash.zero (* place holder *);
-    }
+    {shell_header with proto_level; fitness}
   in
   (if Protocol_hash.equal protocol pred_protocol then return (context, message)
   else