+++ src/lib_shell/prevalidation.mli
@@ -66,14 +66,6 @@ module type T = sig
 
   val apply_operation : t -> operation -> result Lwt.t
 
-  type status = {
-    applied_operations : (operation * Proto.operation_receipt) list;
-    block_result : Tezos_protocol_environment.validation_result;
-    block_metadata : Proto.block_header_metadata;
-  }
-
-  val status : t -> status tzresult Lwt.t
-
   val validation_state : t -> Proto.validation_state
 
   val pp_result : Format.formatter -> result -> unit